Move tedge_config under providers
- There can be many configuration providers, not just TEdgeConfig.
