+++ tedge_config/src/lib.rs
@@ -1,13 +1,13 @@
 mod config_setting;
 mod error;
 mod settings;
-mod tedge_config;
+mod providers;
 mod traits;
 mod types;
 
 pub use self::config_setting::*;
 pub use self::error::*;
 pub use self::settings::*;
-pub use self::tedge_config::*;
+pub use self::providers::tedge_config::*;
 pub use self::traits::*;
 pub use self::types::*;
