+++ tedge_config/src/providers/mod.rs
@@ -1,0 +1,1 @@
+pub mod tedge_config;
