+++ tedge_config/src/providers/tedge_config/data.rs
